feature-extraction-backend-main/pyproject.toml
[project]
name = "feature-extraction-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "jsonschema>=4.23.0",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.35",
]
